+++ package.json
@@ -58,7 +58,6 @@
     "@nextcloud/moment": "^1.3.1",
     "@nextcloud/router": "^3.0.1",
     "@nextcloud/vue": "^8.16.0",
-    "@nextcloud/vue-dashboard": "^2.0.1",
     "autosize": "^6.0.1",
     "color-convert": "^2.0.1",
     "color-string": "^1.9.1",
